requirements.txt
asgiref==3.3.1
autopep8==1.5.4
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
Django==3.1.7
django-environ==0.4.5
djangorestframework==3.12.4
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.11
jupyter-core==4.7.1
jwcrypto==0.8
kiwisolver==1.3.1
matplotlib==3.4.0
numpy==1.19.5
oauth2client==4.1.3
pandas==1.2.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.16
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.8.0
pynput==1.7.3
pyparsing==2.4.7
Pyrebase4==4.4.3
python-dateutil==2.8.1
python-jwt==3.3.0
pytz==2020.5
pywin32==300
pyzmq==22.0.3
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
